+++ src/app/components/info-page/info-components/intro-page/intro-page.component.html
@@ -1,90 +1,14 @@
-<mat-card-title>yes</mat-card-title>
-<mat-card-content>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-  test <br>
-</mat-card-content>
+<div class="info-page">
+  <mat-card-header>
+    <mat-card-title>
+      Hello there!
+    </mat-card-title>
+  </mat-card-header>
+  <mat-card-content>
+    <div class="intro-text">
+      <p>This is the boring info page where I'm supposed to tell you what you can do here.</p>
+      <p>Instead, I'll just let you click around and find interesting stuff</p>
+      <p>...or you can just go back up and look at the pretty floaty dots?</p>
+    </div>
+  </mat-card-content>
+</div>
