+++ src/app/components/info-page/info-components/intro-page/intro-page.component.scss
@@ -1,0 +1,14 @@
+.info-page {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+mat-card-content {
+  padding: 20px;
+}
+
+.intro-text {
+  padding: 10px;
+}
