+++ src/app/components/info-page/info-page.component.html
@@ -4,9 +4,38 @@
       <mat-icon class="info-page-icon">keyboard_arrow_up</mat-icon>
     </button>
   </div>
+
   <mat-card>
-    <div class="inner-card">
-      <router-outlet></router-outlet>
-    </div>
+    <mat-card-header>
+      <button mat-icon-button [matMenuTriggerFor]="cardMenu" class="menu-button">
+        <mat-icon>more_vert</mat-icon>
+      </button>
+      <span class="toolbar-spacer"></span>
+
+      <div class="info-link-buttons">
+        <button mat-icon-button (click)="externalLinkService.openCV()">
+          <mat-icon svgIcon="cv"></mat-icon>
+        </button>
+        <button mat-icon-button (click)="externalLinkService.openGithub()">
+          <mat-icon svgIcon="github"></mat-icon>
+        </button>
+        <button mat-icon-button (click)="externalLinkService.openLinkedIn()">
+          <mat-icon class="icon" svgIcon="linkedin"></mat-icon>
+        </button>
+      </div>
+    </mat-card-header>
+
+    <mat-card-content>
+      <div class="inner-card">
+        <router-outlet></router-outlet>
+      </div>
+    </mat-card-content>
+
   </mat-card>
 </div>
+
+<mat-menu #cardMenu='matMenu'>
+  <button mat-menu-item>
+    <span>ass</span>
+  </button>
+</mat-menu>
