+++ src/app/components/info-page/info-page.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { PanelScrollService } from 'src/app/services/panel-scroll.service';
+import { ExternalLinkService } from 'src/app/services/external-link.service';
 
 @Component({
   selector: 'app-info-page',
@@ -9,7 +10,8 @@ import { PanelScrollService } from 'src/app/services/panel-scroll.service';
 export class InfoPageComponent implements OnInit {
 
   constructor(
-    public panelScrollService: PanelScrollService
+    public panelScrollService: PanelScrollService,
+    public externalLinkService: ExternalLinkService
   ) { }
 
   ngOnInit() {
