+++ src/app/components/splash-page/splash-page.component.ts
@@ -3,6 +3,7 @@ import { ThemeService } from 'src/app/services/theme.service';
 import anime from 'animejs/lib/anime.es.js';
 import { PanelScrollService } from 'src/app/services/panel-scroll.service';
 import { ConfigurationService } from 'src/app/services/configuration.service';
+import { ExternalLinkService } from 'src/app/services/external-link.service';
 
 @Component({
   selector: 'app-splash-page',
@@ -20,7 +21,7 @@ export class SplashPageComponent implements OnInit, AfterViewInit {
   constructor(
     public themeService: ThemeService,
     public panelScrollService: PanelScrollService,
-    private configService: ConfigurationService
+    public externalLinkService: ExternalLinkService
   ) { }
 
   ngOnInit() {
@@ -38,27 +39,6 @@ export class SplashPageComponent implements OnInit, AfterViewInit {
     });
 
     this.iconButtons = document.querySelectorAll('button.icon-button');
-  }
-
-  /**
-   * Function that opens the CV
-   */
-  public openCV() {
-    window.open(this.configService.config.externalLinks.cv, '_blank');
-  }
-
-  /**
-   * Function that opens the CV
-   */
-  public openGithub() {
-    window.open(this.configService.config.externalLinks.github, '_blank');
-  }
-
-  /**
-   * Function that opens the CV
-   */
-  public openLinkedIn() {
-    window.open(this.configService.config.externalLinks.linkedin, '_blank');
   }
 
   /**
